+++ lib/github_signin_promax.dart
@@ -2,3 +2,4 @@ library github_signin_promax;
 
 export 'package:github_signin_promax/src/signin_screen.dart';
 export 'package:github_signin_promax/src/signin_params.dart';
+export 'package:github_signin_promax/src/signin_response.dart';