+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 0.0.4
+- Change enum for Response status
+- handle when user cancel by press back button
+- Update Readme and example